FEAT : 방 기능 추가 구현

+++ src/room/room-service.ts
@@ -6,7 +6,13 @@ import { AppDataSource } from '../../config/db/data-source.js';
 
 export const createRoomService = async (managerId: string, title: string, description: string, playListUrl: string, playList: any[]) => {
   const roomRepository = AppDataSource.getRepository(Room);
-  const thumbnail = await getPlaylistThumbnail(playListUrl);
+  let thumbnail;
+
+  try {
+    thumbnail = await getPlaylistThumbnail(playListUrl);
+  } catch (error) {
+    throw new Error('Invalid URL');
+  }
 
   const newRoom = roomRepository.create({
     managerId,
@@ -15,14 +21,15 @@ export const createRoomService = async (managerId: string, title: string, descri
     playListUrl,
     roomThumbnail: thumbnail,
     playList,
-    roomMember: [{ userId: managerId, nickName: '관리자' }], // 방 생성자는 자동으로 관리자 역할로 추가
+    roomMember: [{ userId: managerId, nickName: '관리자' }], 
     chats: [],
   });
 
   await roomRepository.save(newRoom);
 
-  return { success: true, message: 'Created' };
+  return { success: true, message: '방이 생성되었습니다.' };
 };
+
 export const getRoomsService = async () => {
   const roomRepository = AppDataSource.getRepository(Room);
   const rooms = await roomRepository.find({
@@ -39,13 +46,19 @@ export const getRoomsService = async () => {
   return { roomList };
 };
 
-export const getRoomService = async (roomId: string) => {
+export const getRoomService = async (roomId: string, userId: string) => {
   const roomRepository = AppDataSource.getRepository(Room);
   const objectId = new ObjectId(roomId);
-  const room = await roomRepository.findOne(objectId as FindOneOptions<Room>);
-
+  const room = await roomRepository.findOne({ where: { _id: objectId } });
+  
   if (!room) {
-    return null;
+    return { success: false, message: '방을 찾을 수 없습니다.' };
+  }
+
+  const isMember = room.roomMember.some(member => member.userId === userId);
+
+  if (!isMember) {
+    return { success: false, message: '방에 참여하지 않은 사용자입니다.' };
   }
 
   const { roomTitle, roomDescription, roomThumbnail, playList, roomMember, chats } = room;
@@ -61,42 +74,37 @@ export const getRoomService = async (roomId: string) => {
   };
 };
 
-export const deleteRoomService = async (roomId: string) => {
+export const deleteRoomService = async (roomId: string, userId: string) => {
   const roomRepository = AppDataSource.getRepository(Room);
   const objectId = new ObjectId(roomId);
-  const room = await roomRepository.findOne(objectId as FindOneOptions<Room>);
+  const room = await roomRepository.findOne({ where: { _id: objectId } });
 
   if (!room) {
-    return { error: 'Room not found' };
+    return { success: false, message: '방을 찾을 수 없습니다' };
+  }
+
+  if (room.managerId !== userId) {
+    return { success: false, message: '권한이 없습니다' };
   }
 
   await roomRepository.remove(room);
-  return { success: true, message: 'Deleted' };
+  return { success: true, message: '방이 삭제되었습니다' };
 };
 
 export const joinRoomService = async (roomId: string, userId: string, nickName: string) => {
   const roomRepository = AppDataSource.getRepository(Room);
   const objectId = new ObjectId(roomId);
-  const room = await roomRepository.findOne(objectId as FindOneOptions<Room>);
+  const room = await roomRepository.findOne({ where: { _id: objectId } });
 
   if (!room) {
-    return { error: 'Room not found' };
+    return {success: false,  message: '방을 찾을 수 없습니다.' };
   }
 
-  // 이미 방에 참가한 사용자인지 확인
   const isAlreadyMember = room.roomMember.some(member => member.userId === userId);
   if (!isAlreadyMember) {
     room.roomMember.push({ userId, nickName });
     await roomRepository.save(room);
   }
 
-  const { roomTitle, roomDescription, roomThumbnail, playList, roomMember, chats } = room;
-
-  return {
-    nickName,
-    userId,
-    playList,
-    roomMember,
-    chats,
-  };
+  return { success: true, message: '성공적으로 방에 가입했습니다.' };
 };
